Rename: `ReaderJsWrapper` -> `ReaderJsHelper`
